add validation check
